# contracts/control-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: AstrLink Local Control API
  version: 1.0.0
  summary: Stable local contract between the AstrLink desktop application and Core.
  description: |
    This API is a local control plane; it is not part of the public inference API.
    Core binds the inference plane to `127.0.0.1:18317` by default and binds this
    control plane to the ephemeral address `127.0.0.1:0`. On startup Core writes
    exactly one ready JSON line to stdout containing the selected `control_url`.
    Launchers must use that value rather than guessing a port.

    The bootstrap health, version, and capability operations are intentionally
    unauthenticated and disclose no configuration. Every operation that reads or
    mutates user configuration, events, or request records requires the per-start
    local control token. The token must be delivered through an inherited pipe or
    equivalent protected IPC; it must not be placed in process arguments or the
    environment.

    Authenticated callers hold one of two roles, recorded on each operation as
    `x-astrlink-role`. The desktop's per-start control token is `operator` and may
    call everything. Agent tools reach Core through the same-user control socket or,
    on Windows, a separate per-start observer token; both are `observer`, which may
    call only GET operations marked `observer`. Every non-GET operation, token
    reveal, and OAuth authorization session is `operator`, and an observer calling
...
    operations.

...

...
servers:
...
security:
  - controlToken: []
...
paths:
...
  /control/v1/observers:
    x-astrlink-implementation-status: implemented
    get:
      x-astrlink-role: observer
      tags: [Requests]
      operationId: getObservers
      summary: Report recent agent-side reads of the control plane.
      description: |
        Requests from the local control socket or the MCP bridge count as agent
        reads. The desktop uses this to show that an agent is watching and how far
        its latest read reached. The state lives in Core memory only.
      responses:
        "200":
          description: Observer state.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObserversResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    controlToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-control-token
      description: Per-start local token delivered over protected IPC; never persist it in configuration or logs.
...
  headers:
    NoStore:
      description: Sensitive local-control responses must not be cached.
      schema:
        type: string
        examples: [no-store]
...
  responses:
...
    Unauthorized:
      description: Missing or invalid local control token.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    RawAccessDecisionMode:
      type: string
      enum: [once, window_15m, deny]
...
    RawAccessEvent:
      type: object
      required: [at, kind, client_name]
      description: |
        One step of an agent raw access request. Identifiers only; never content
        or proof. A `password_invalid` event from the desktop's own unlock or
        password change carries no `grant_id` or `request_id`.
      properties:
        at:
          type: string
          format: date-time
        kind:
          type: string
          enum: [requested, approved, denied, password_invalid, raw_read]
        grant_id:
          type: string
          pattern: "^rawgrant_[0-9a-f]{16}$"
        request_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        client_name:
          type: string
          maxLength: 64
        decision:
          $ref: "#/components/schemas/RawAccessDecisionMode"
      additionalProperties: false
...
    ObserversResponse:
      type: object
      required:
        - last_seen_at
        - client
        - requests
        - read_level
        - last_raw_read_at
        - pending_raw_access
        - raw_access_events
      properties:
        last_seen_at:
          type: [string, "null"]
          format: date-time
        client:
          type: string
          maxLength: 64
        requests:
          type: integer
          minimum: 0
        read_level:
          type: string
          enum: ["", metadata, shareable, raw]
          description: Level of the latest agent-side request; empty before any.
        last_raw_read_at:
          type: [string, "null"]
          format: date-time
        pending_raw_access:
          type: integer
          minimum: 0
        raw_access_events:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/RawAccessEvent"
          description: Recent raw access steps, oldest first.
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/ControlError"
        request_id:
          type: string
          minLength: 1
          maxLength: 96
      additionalProperties: false
    ControlError:
      type: object
      required: [code, message, retryable, details]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 96
          pattern: "^[a-z][a-z0-9_]*$"
          examples: [missing_protocol_capability, validation_failed]
        message:
          type: string
          minLength: 1
          maxLength: 1024
          description: Sanitized user-facing message.
        retryable:
          type: boolean
        details:
          type: array
          maxItems: 32
          items:
            $ref: "#/components/schemas/ErrorDetail"
      additionalProperties: false
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          maxLength: 256
        reason:
          type: string
          maxLength: 512
        protocol:
          $ref: ./protocol-capabilities.schema.json#/$defs/ProtocolId
        service_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        required_plan_types:
          type: array
          uniqueItems: true
          items:
            $ref: ./protocol-capabilities.schema.json#/$defs/PlanType
        retry_after_seconds:
          type: integer
          minimum: 1
      additionalProperties: false
...